Reorganize the source files (fixes #79)
