+++ src/precompile.jl
@@ -32,7 +32,7 @@ function _precompile_()
     precompile(Tuple{typeof(prepare_call), Any, Vector{Any}})
     precompile(Tuple{typeof(Core.kwfunc(prepare_call)), NamedTuple{(:enter_generated,),Tuple{Bool}}, typeof(prepare_call), Function, Vector{Any}})
     precompile(Tuple{typeof(Core.kwfunc(prepare_framecode)), NamedTuple{(:enter_generated,),Tuple{Bool}}, typeof(prepare_framecode), Method, Type})
-    precompile(Tuple{typeof(build_frame), JuliaFrameCode, Vector{Any}, Core.SimpleVector})
+    precompile(Tuple{typeof(prepare_frame), JuliaFrameCode, Vector{Any}, Core.SimpleVector})
     precompile(Tuple{typeof(extract_args), Module, Expr})
     precompile(Tuple{typeof(enter_call), Int, Int})
     precompile(Tuple{typeof(enter_call_expr), Expr})
